docker: update docker/checks from core, add cmake
problem: sched's containers, matrix and check scripts were out of date
compared to core

solution: Pull in the new bookworm and fedora38 containers and multiarch
build setup from core, along with updated check files and fixes. We now
get arm64 containers out the other end.  I tried getting 32-bit also,
but it is very, very broken, and also complicated by a bug in qemu-user
and/or the linux kernel depending on who you ask.  Would be really nice
to have that extra check, but it's got to be a different PR at this
point. Add cmake and ninja to all containers for the new build setup.
While not strictly required, ninja is tiny and makes the builds and
rebuilds faster when used.