add cmake build system alongside autotools
This is a big one, that was originally developed as a series of commits,
major decisions are summarized below, but the short version is we have a
cmake-based build system that can be configured and built as though it
still had autotools.  To use the autotools build system, manually run
autoreconf to replace the configure script, and then build as usual.

We're adding a fake configure script here so that it has approximately
the same ergonomics as the autotools build system.  The only thing that
causes some breakage is that the autogen.sh script is a no-op, so
autoreconf needs to be run manually to do the autotools side of the
build system.

To mimic the autotools build system:
1. respect PYTHON env var

  This is a QOL improvement to allow specifying the python version the
  autotools way rather than with the cmake defines.

2. find flux-core by finding flux, match prefix if none specified

  @grondo pointed me to a feature I didn't know we had, where if
  flux-core is in path it's automatically found.  This does that in the
  cmake build system both for the way autotools does it, by checking the
  install prefix, and by checking path as well.

Note on yaml-cpp, while they provide an upstream cmake config that I
would prefer to use, not all of our docker containers actually offer
that support, so we're still using pkg-config for now.