testsuite: improve test event check
Problem: when building against the installed flux-core on
a system with flux-coral2 installed, one of the 'make check'
tests fails.

A grep for "finish" in a job eventlog succeeds where it failed
before because the cray-pals port distributor jobtap plugin
posts a prolog-finish event.

Aside: the cray-pals jobtap plugin is loaded by an rc script
rather than the config file so that it is present in all instances,
not just the system instance.

Use 'flux job wait-event finish' instead of dumping the human
formatted eventlog to a file and grepping it.  Note that wait-event
exits immediately with an error when the eventlog is complete but
the requested event is missing.

Fixes #1073