build: support 'make deb'
Problem: The 'make deb' target was dropped in the conversion to cmake,
but having an easy way to build test .deb packages has become an
essential component of testing.

Add a custom 'deb' target to CmakeLists.txt. Since this target requires
'make dist', add a 'dist' custom target as well.

Some versions of git do not support the `--add-virtual-file` option,
so instead of using that to add the flux-sched.ver file to the archive,
just create the file, add it, and subsequently remove it.

Fixes #1072