build: install libdir/python3.9/fluxion link
Problem: In the previous autotools based build system, a soft link from
libdir/python<ver>/fluxion was created to point to the real fluxion
directory in python site packages. This was done because the flux(1)
command driver actually prepends this non-standard PATH to PYTHONPATH
in order to prepend just the path to the Flux modules instead of the
path to all site-installed modules. In the conversion to cmake this
link was dropped from `make install`.

Re-establish this python link in `make install`.

Fixes #1087