build: do not install libresource.so
Problem: libresource.so was previously a convenience library, but is
now installed by default. However, it has a generic name that could
easily become confusing, and is named simply `libresource.so`, which
doesn't follow standard shared library naming conventions.

Make libresource a static convenience library instead and do
not install it.  The library should be renamed something like
`libfluxion-resoruce.so` and possibly be versioned with an soname if
installed in the future.

Fixes #1093