build: do not build html docs by default
Problem: In the previous version of the build system, html
documentation was only built in ReadTheDocs or explicitly via `make
html`, but currently html docs are built by default. This causes
the build to fail if the sphinx_rtd_theme is not installed, whereas
previously the build would succeed (with manpages available).

Remove html from the default docs build target.