ci: do not force configure --prefix to /usr
Problem: When flux-sched is built in ci using the configure wrapper
for cmake, a prefix of /usr is forced. This bypasses the automatic
setting of libdir to match flux-core and results in Fluxion modules
installed in the wrong path.

Drop the use of --prefix=/usr in generate-matrix.py. This isn't
necessary since the prefix will be set to that of the first flux(1)
found in PATH, which is /usr/bin/flux.

Fixes #1099