reapi: track match success and update info for match allocate
Problem: reapi_cli_t::match_allocate does not return or track
whether a match was successful or not. It also sets the job info
with an unset value for overhead.

Check return code from traverser run to determine if a match is
successful. If a match is not successful, increase the job
counter and return 0. errno is set with the correct error value
to be handled by the calling function.

Move code blocks that set overhead value ov above blocks where
job info is created.