build: force libdir to match that of detected flux-core
Problem: When building with an explicit prefix, the build system will
derive a libdir from that prefix which may or may not match flux-core.
Since Fluxion modules are installed under the derived libdir, they
may not end up in the default module search path.

Always use the libdir from flux-core when configuring to ensure that
Fluxion modules are in the default search path.