tests: update golang tests and output
Problem: a recent fix to reapi changed the output of the golang
bindings tests. This changed a value that was once static to a
value that changes with every run.

Use a sed command to remove this value from the output of the test.
Edited tests to use orelse_reserve flag.