ci: fix coverage build
Problem: It appears that the coverage check commands return a
nonzero exit status even when `make check-code-coverage` succeeds,
and  therefore `make recheck` is unnecessarily executed.

The failure appears to be in one of the post-check steps,
which are possibly duplicated from the cmake version of `make
check-code-coverage`.  In any event, remove 'coverage html' (which
probably isn't useful anyway), and make success of `coverage xml`
and `coverage report` optional.