traverser: fix units of graph_end in traverser::run()
Problem: In dfu_traverser_t::run() the current graph_end is compared
to the current `*at` time to ensure the job duration won't overrun
the graph expiration. However, the local graph_end variable is created
using the default graph_duration.graph_end internal count, which may
not be in seconds.

Explicitly cast the graph_end to std::chrono::seconds so that the
value of the local graph_end variable is the same units as `*at` and
`duration` which are seconds.

Fixes #1103