reapi: simplify build entries for go tests
Problem: the Go test links against unnecessary libraries such as Fluxion
resource, reapi_module, and flux::core. Furthermore, the options are a
long, unwrapped string which is hard to read.

Remove the unnecessary entries in the CGO_LIBRARY_FLAGS and make the
presentation more readable. Update GolangSimple.cmake to remove string
replacement of " " with ";" which causes the list expansion to fail.