python: add 'status' to ResourcePool constructor
Problem: the FluxionResourcePoolV1 does not support setting
vertex status. However, some resources, such as rabbits, should
be initialized as down, and are only known to Fluxion through
JGF produced by the FluxResourcePool class.

Add a 'status' parameter to the FluxionResourcePoolV1 constructor,
with a default of 'up' to maintain backwards compatibility.