reader: scrub JGF reader fetcher status value
Problem: when the JGF reader encounters a vertex with its status
specified in the JGF, the fetcher sets its status appropriately but
does not scrub the value. This causes all subsequently fetched vertices
to inherit the value.

Add a line to reset the status to the JGF default value in scrub ().