planner_multi: use boost multi_index, update c interface
Problem: to update the Fluxion resource graph, the traverser will need
to be reinitialized, which assembles vectors of resource counts and
types. The ordering of the counts and types depends on how graph is
mutated. Currently, counts and types are stored as separate vectors in
`planner_multi` under the assumption that their order and size don't
change. Removing that assumption will cause inconsistency in
scheduling.

Add a `boost::multi_index` container to replace and simplify
`planner_multi` member data. The `boost::multi_index` container has the
important capability of searching and comparing by multiple indices
which can be vector-like (`random_access`), and hash-like
(`hashed_unique`). Adding two indexes for lookups enables support for
altering resource types, counts, and their orders (indices) in a
natural way. Performing these transformations with `vectors` would be
high-complexity both in time and code.