debbuild: fix Python purelib paths on Debian systems
Problem: Debian based systems may use the 'posix_local' scheme for
Python installations, which causes `make deb` to break because the
'purelib' path is returned under /usr/local even when the flux-core
prefix is /usr.

Use the workaround described in

 https://gitlab.kitware.com/cmake/cmake/-/issues/25113

and set DEB_PYTHON_INSTALL_LAYOUT=deb to fix `make deb`.

There is probably some standard way to actually fix this issue,
but since the deb builds are purely for testing at this time, this
solution works for now.