reader: check type of properties in jgf
Problem: the JGF reader requires each vertex's 'properties' field
to be either null or a JSON object, but doesn't ever check the type.
This leads to silent failures when the type read in is wrong.

Add a check for the type of the properties field if it is not null.