planner interface: add span reduction functionality
Problem: the planner currently cannot remove resources from a span. This
functionality is necessary for partial cancellation.

Add a helper function to reduce the span scheduled quantity. Add an
interface function to remove span resources and check if the values to
be removed are valid. Special care must be taken for 0 spans or removal
of 0 resources. Zero spans are used to track individaul planner
resource spans as part of a span of multiple planner resource types
tracked by a planner_multi. Add a check for removing 0 resources which
can also be used to check if a span is valid. Return an updated bool
which indicates whether all span resource have been removed.