fix: convert strings to boost:flyweight
Problem: string comparison is immensely inefficient, taking 6-10%
of resources (reported by Tom) for a trace.
Solution: start with the root of the issue in the scoring
module and work backwards to convert string types to boost::
flyweight. I tried to have a minimal footprint here but it
spread really quickly

Signed-off-by: vsoch <vsoch@users.noreply.github.com>