qmanager: split remove to separate pending
problem: The "remove" method was used for both handling the `free` and
`cancel` RPCs. Under some code paths, remove de-allocates resources, so
using it for a pending cancellation, even if it seems to be protected by
checks, is unnecessarily dangerous.  Given the recent double-booking
issue and a lack so far of other theories on how we get there, this is a
possible culprit.

solution: Re-factor queue_policy_base_t remove into remove and
remove_pending, the latter of which only operates on the pending states
and has no code-paths that can result in a resource.cancel RPC being
sent.  The remove_pending is used for both the cancel RPC and the part
of the free RPC where the code was used before, should be a pure
refactor.