rv1exec reader: implement update ()
Problem: issue https://github.com/flux-framework/flux-sched/issues/991 identified the need for an `rv1exec`
implementation of update (). The need for the implementation is
described in detail in the issue, but the primary motivation is to
enable reloading Fluxion when using RV1 without the scheduling key and
payload. The reader was not originally implemented due to the lack of
information in the format. Examples include edges, exclusivity, paths
with subsystems, and vertex sizes. To create a workable implementation,
strong assumptions need to be made about resource exclusivity and
edges.

Add support for update () through helper functions that update vertex
and edge metadata and infer exclusivity of node-level resources.