testsuite: update existing tests, add checks for update under rv1exec
Problem: the current tests for reloading Fluxion with rv1exec fail due
to lack of update support. Now that the support is added, we need tests
for correct behavior.

Add new tests to check for ability to reload Fluxion and update a
resource graph with rv1exec. Update existing tests to use the
additional option required by resource-query for update.