qmanager: look up job queue name in KVS
Problem: when the fluxion modules are reloaded with running jobs,
and the match format is "rv1_nosched", and queues are enabled,
running jobs are killed with a fatal scheduler-restart exception.

During the hello handshake defined by RFC 27, the scheduler is informed
during its initialization of jobs that are holding resources.  The hello
callback in qmanager retrieves the job's queue name from the R key
"attributes.system.scheduler.queue".  The queue name is used to locate
the proper queue for the job to be inserted into.

This attribute is not being set in R when the match format is
"rv1_nosched" so the default queue is assumed.  Since the default queue
is not instantiated when named queues are defined, a fatal job exception
is raised when the queue lookup fails.

There has been a proposal to deprecate the R attribute (#1108), so
rather than ensure that it is set in this case, determine the queue
instead by fetching the jobspec from the KVS.

Fixes #1108