qmanager: fix error handling during hello
Problem: several error paths in jobmanager_hello_cb() do not
result in the function returning a failure.

I think the only ones that do trigger an error are a C++ exception
(due to the wrapper) and if (queue->reconstruct() returns -1. The
callback returns 0 (success) in other cases which could result in
double booked resources.