cmake: add support for sanitizers
problem: using address or other sanitizers is currently very manual and
error prone

solution: add support into the build system to select appropriate
sanitizers and link them in.  This uses an external project which has
been pulled in as a git subtree (NOT a submodule, it's a single squashed
out commit that pulls in the directory).