cmake: add presets file to ease use of asan
problem: build configurations can be complicated to manage, especially
for sanitizers and similar.

solution: cmake supports a CMakePresets.json file to define commonly
used build and test configurations.  Add one of these that has the
correct configuration for configuring, building and testing both in the
default mode and with address sanitizer.