resource: fix decode of rankset="all"
Problem: When Fluxion marks "all" resources down, it uses the size of
the by_rank map to encode the set of all ranks, but this is incorrect
because excluded ranks are never presented to the scheduler by the
flux-core resource module. This can result in down resources being
scheduled by Fluxion, causing jobs to immediately fail.

Modify decode_all() to use the actual ranks in the by_rank mapping
instead of presuming the ranks are 0-(size-1).

Fixes #1182