cmake: write flux-sched.ver to build directory
problem: dist was modifying the source directory

solution: write the file to the build directory and rely on the basename
behavior of add-file to add it to the right place

The original version of this used `--add-virtual-file`, which should
have taken care of it in a simpler way, but it seems that flag is
ignoring the `--prefix` argument at least on my system.  Plan to
investigate further and report upstream if it remains an issue.