resource: declare extern stats and add max-match data
Problem: the resource module and traverser need to declare the external
`perf` struct. Also, tracking the jobid and `schedule` iternation count
of the slowest match time has become important for diagnosing match and
queue performance degradation.

Add the extern declaration of the `perf` struct and add capability to
track the jobid of the slowest match time and the respective number of
schedule iterations.