resource_graph: avoid copying std::map in out_edges
problem: out_edges is called for every single vertex in every traversal,
and every time it's called the subsystem_selector_t is copied.  The cost
of this was approaching 50% in some perf traces because a std::map was
being stored in the structure, despite the fact a copy of the map is
owned by the matcher context and could be reused.

solution: store a const pointer to the long-lived original