qmanager: search blocked jobs during pending removal
problem: if a job is pending and moved into the blocked state, it isn't
removed by cancel

solution: if the job is not found in pending, try to remove it from
blocked

As a note, I remember writing this logic at some point in the previous
PR that added m_blocked.  I don't know when it got lost, but probably
when a few other similar searches that we actually don't want got
removed.  The job is now cancelled appropriately in this case at least.

fixes #1208