readers: add experimental partial cancel support for JGF
Problem: JGF partial cancellation is more complex and flexible than
RV1exec. Unlike RV1exec, there is no top-level metadata that tracks the
ranks to be canceled. That means the full JGF string needs to be
unpacked, and each vertex fetched and cancelled individually.

Add the capability to cancel each vertex in the JGF subgraph and track
the number and types of resources to be removed to pass back to the
traverser for its top-down removal.

This functionality is currently experimental and will support cloud
deployments. The implementation can be extended in a straightforward
way to arbitrary resource release.