qmanager_callbacks: run sched loop even if in progress
problem: Currently the sched loop only gets poked if both the queue is
schedulable _and_ the sched loop is not in progress.  I need to break up
the scheduling loop of backfill schedulers to allow the reactor to make
progress, so it needs to call into the sched loop whenever it is
schedulable.

solution: remove the requirement that the loop not be in progress

This is an atomic commit on the way to something bigger, but it's here
because this change is correct by itself.  None of our tests change
correctness or performance with this change.