qmanager/bf: process the scheduling loop in chunks
problem: qmanager blocks for the entire duration of the scheduling loop,
which prevents message processing and holds up many other processes like
submitting new jobs

solution: rework the backfill policy base to store the loop progress in
members of the policy class, and if it needs to be called again returns
-1 with errno=EAGAIN.  I'm not super fond of this as an interface
design, but it fits with the current interface in place.  This way, each
time run_sched_loop is called, it will process N (which is currently 5)
match attempts before setting itself as schedulable and returning to the
reactor to process new messages.  This way the maximum delay between
allowing the reactor to make progress is limited.

I would still like to make this asynchronous, but doing that without
changing the algorithm is going to take more work given the way messages
between qmanager and resource are set up.