formatting: add clang-format file
problem: we end up wasting a lot of time on manual reformatting

solution: encode our format in clang-format so we can not do that.

There is more work to be done with this, including applying everywhere
and deciding what we do with jansson formatting, but I'm using it in
this PR so it seems appropriate to at least have it here.