t/unreservable: add head-of-line blocking test
problem: We need a test to ensure we can schedule jobs past the queue
depth in the presence of enough blocked jobs

solution: Add new test clauses to the unreservable test to ensure this
case can pass.  Also I'm removing the limit on the number of failed
considerations, it's too fragile to keep in, and it's effecitvely a
performance test we should be handling a different way.