qmanager/fcfs: ensure post_sched_loop is invoked
problem: fcfs relies on post_sched_loop running while the scheduling
loop is active to handle priorities correctly, and possibly to make
progress

solution: when a match failure is received from match_multi, explicitly
set m_scheduled to true so that the prep watcher will fire the check
watcher to run the post_sched_loop after the queue is no longer
processing the sched loop.  This is necessary to allow us to protect the
post_sched_loop from running during a scheduling loop in backfill
schedulers.

Goodness help me this is a lot of complex interactions and flags, and it
needs a refactor really, really badly. That said, this seems to check
every box I can think of.