t/priority: test freeing blocked jobs with reprio
problem: at least in conservative, if a job with a higher priority has
its priority lowered, it should reconsider blocked jobs

solution: add a test to t1013-qmanager-priority.t. Specifically we use
the conservative policy because the problem doesn't seem to manifest as
easily with "easy", but this test _does_ fail without the
`reconsider_blocked_jobs` calls in reprio.