qmanager/backfill: make backfill queues asynchronous
problem: blocking RPCs with resource, and a blocking scheduling loop,
cause severe delays when a large volume of messages is received. Even
after making the sched loop itself asynchronous and periodically
returning to the reactor, the ordering and delays caused by the
synchronous RPCs made processing large volumes of message prohibitively
slow, @milroy reported as much as 4422 seconds for 10,000 allocs and
cancels in quick succession.

solution: use chained calls to match_allocate_multi with one job each to
implement a fully asynchronous scheduling loop interaction with
resource.  This does not implement sched loop cancellation, which would
probably also be good, but even without it does the same 10,000 allocs
and cancels in roughly 45 seconds on my laptop for almost a 100x speedup
and an even larger improvement in overall responsiveness.