qmanager/bf: ensure efficient forward progress
problem: with the spinning removed, scheduling would proceed to the end
of the queue depth and then stop.  This is as designed, _but_ it is
expected that the `schedulable` flag is set when a new event comes in,
and not cleared by `run_sched_loop`, it's also expected that there will
be periodic events to trigger this.  The rest of this PR removed all of
those cases, so we would now finish one sched loop, then park until
something new came in or finished, essentially locking ourselves.

solution: Two changes are required to make this deterministically make
progress, 1) only unset schedulable when _actually starting a scheduling
loop_, meaning when the loop is not currently active, 2) when the loop
is complete and setting active to false, if it is ending because it has
reached the end of the queue depth, mark schedulable to start a new
loop.