dfu: add a quick-to-test feasibility precheck to Match
problem: When a large number of possible time points exist, unreservable
jobs cause N matches where N is the number of time points. This makes
jobs that can't be matched potentially very expensive, and heavy use of
constraints make the case common.

solution: iterate over just the node vertices checking for constraint
matches, whether they are UP and whether it's possible to allocate them
(either at the end of time, or in the case of allocate immediately). If
all of those pass, the node is deemed "feasible" and counted. If less
nodes are feasible than are required by the request, it's immediately
rejected.

In a test with 300 time points, without this patch I get between 5 and
7.5 seconds for each failing match. With this patch the failed match
stats in the same test report a minimum of 0.005 and a maximum of 0.008
seconds. This will not help with *successful* matches, but for matches
that can't work because of drained nodes, constraints, queues, etc. it
should reduce match times massively.