planner/dfu: avoid dereferencing empty vector
problem: a few places were using `&(vec[0])` to get an address of the
start of a potentially empty vector.  This is almost always actually
safe in practice, but it's not allowed and technically could be treated
like a null dereference as well.

solution: use the `vec.data()` method that exists specifically for this
purpose