graph: use builtin bidirectional graph
problem: walking inward edges is difficult and slow because they're
mixed in with the `out_edges` range (we explicitly add out-edges to
every vertex matching each inward edge for some reason).  This means we
have to filter them out during traversals, and maintain them manually.
They show up almost like another subsystem (the "in" subsystem) in some
readers, and aren't generated at all by others.

solution: switch from directed to bidirectional graph type, remove the
explicit addition of "in" edges, and fix the graph test to match the
actual number of edges.

I was very surprised how easy this was, would have sworn I'd tried it
and failed at least once before.  The main possible downside is that
bidirectional graphs have to store every edge twice, and thus should
require twice as much storage for edges as directional graphs.
Strangely, switching to this actually didn't increase memory usage _at
all_.  It stayed identical, and I don't know why.  The even _better_
news is that after removing the explicit back-edges, this actually saved
500MB for a 16,000 node 64-core graph, and made running the test under
heaptrack take a full second less with 5,000,000 less allocations (17.5s
-> 16.5s and 75m allocations to 70m for just loading the graph and
stopping).

In the end, the only required logic change was in the jgf reader, where
it failed to find a parent vertex because it was only checking out edges
of the child. That is fixed in a simplistic way right now, but could be
optimized massively.