util: add string interner and interned_string class
problem: we store a _lot_ of duplicate strings, especially types and
basenames and subsystems, we also compare them a lot, to the tune of up
to 60% of execution time in some extreme cases.

solution: add an easy-to-use string interner and interned string class
so we can not only store these strings once, rather than n times, but
get a stable, selectably-sized integer identifier for each string that
will be densely ordered starting from zero, so we can actually use small
vectors for things indexed by subsystem.

we also add range and container helper types

problem: we use subsystems as the key for a map container pretty
frequently, which is both bad for storage and performance

solution: add a helper container that uses a single vector for storage
but allows lookup of elements based on the interned subsystem along with
automatic extension on `operator[]`