chore: flux ion-resource jobspec argument redundancy
Problem: the flux-ion-resource.py match has several subcommands
that require a jobspec positional argument. Each subparser is
calling the same logic to add it, which is redundant.
Solution: iterate through a list to add the same argument
to all of them, eliminating the redundancy and making it easier
for the developer to read.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>