traverser: don't add pruning filter for leaf vertices
Problem: when ALL is specified for prune-filters of a leaf vertex type,
other leaf vertices maintain a pruning filter for that resource type.
Since a leaf vertex is the terminus of a depth-first search of a
single-subsystem graph, the extra pruning filters increase memory
consumption unnecessarily.

Add a check that will exit the pruning-filter priming function for leaf
vertices.