qmanager: annotate each job only once on reservation
problem: an annotation is sent for each job that enters reserved state
when its predicted start time changes.  That is controlled by when the
`old_at` and `at` times associated with the job in qmanager are
different.  The implementation fails to update `old_at` after this to
prevent repeat messages.  This was somewhat hidden by the fact that the
subsequent updates had little effect, except after recent performance
improvments made it possible for the qmanager to flood the job-manager
with annotations in some cases.

solution: set old_at to at after deciding whether to send the annotation