matrix: add explicit python path for el8
problem: upstream el8 has added a new version of valgrind, which pulls
in python3.11. The older cmake logic on el8 finds the newer python but
fails to find any development dependencies for it and errors out.

solution: force cmake to use the system python3.6, it must be an
absolute path as well for some reason