qmamager callbacks: add support for final free flag
Problem: recently, a large job on a large system was considered
allocated by Fluxion, but was complete and released in flux-core
(https://github.com/flux-framework/flux-core/issues/6179). The proposed
solution was to amend RFC 27 to include an optional "final" boolean
flag in the `.free` RPC. That flag can be used by Fluxion to determine
if there is an allocation state discrepancy between flux-core and
sched.

Add support to unpack the "final" boolean and send it to the qmanager
policy for handling.