traverser: don't return error for missing x_span in full cancellation
Problem: if a partial cancellation does not fully remove an allocation
but the "final" flag is set by the .free RPC a full cancellation is now
run. However, the current traverser check for a missing exclusive span
(x_span) considers this invalid for a full cancellation and returns an
error.

Update the check to only return an error for this condition if the
cancellation is of type VTX_CANCEL.