cmake: check for compiler versions gcc12+ and clang15+
problem: Cmake will accept a compiler as C++20 compliant if it accepts
the flag, but we require more features than the bare minimum, leading to
compilers appearing to be supported that are not.

solution: Explicitly check for gcc12+ or clang15+, if neither are found
error out with a descriptive message which includes a (hopefully
sufficiently scary) workaround to override this check and build with
another compiler.