cmake: nominally bump language version to C++20
problem: there are features available in our supported compiler versions
we currently can't use because we're marking C++20 as our std version
rather than 23.

solution: bump the c++ standard version to 23.  Note that this _does
not_ change the documented required compiler versions, in fact it even
works for one version older than each in and of itself.  It just turns
on a couple of niceties like `string.contains()` that otherwise we
wouldn't have access to.