qmanager-stats: fix jansson refcount issue
problem: The `to_json` function for serializing maps with jansson had a
refcounting bug in it, which wasn't caught by address sanitizer,
valgrind, or testing.  This is what caused the issue @jamescorbett saw
in the buildfarm, but only reliably reproduced on alpine.

solution: Temporary solution is to fix the refcount problem, longer
term solution is to use something that isn't so insanely hard to use
correctly and debug.