pre-commit: switch from pylint and black to ruff
problem: pylint doesn't run under pre-commit properly, and both black
and pylint are slow

solution: switch to ruff, which is very, very fast and implements both
as well as isort, also run it so this will pass (which is why unused
imports are removed here). Also remove the rule from the gh actions
setup so we don't run both.