resource: move feasibility service from qmanager to resource
Problem: There is no reason the feasibility service and feasibility
RPC need to be registered in qmanager, since that module just synchronously
forwards the RPC to the resource 'satisfiability' RPC.

Register the feasibility service directly in the sched-fluxion-resource
module and add a message handler for feasbility.check to be serviced
by satisfiability_request_cb().

The existing sched.feasibility RPC is still handled by qmanager for
backwards compatibility.