Add feasibility module test
Problem: All calls to feasibility.check are in tests for s-f-resource,
not tests for s-f-feasibility.

Move them to a separate test for the feasibility module.

Disallow load-file behavior in feasibility test

Problem: t4014 expects feasibility to load resources from a resource
module that was passed a load-file, which is not desired
behavior.

Update t4014 to expect failure on such a load.