interner: add finalize/open/close to interner
problem: refcounted strings for resource types are turning out to be
expensive enough to actually slow us down.  We compare them and copy
them so often they show up in the traces as 20-30% of runtime even as
just shared_ptrs

solution: avoid the possible denial of service issue from resource types
by adding an interface to the interner type that finalizes it,
preventing new strings from being added until it is explicitly "opened"
for addition.  This way after initialization of the graph/database we
finalize the interners for subsystems and resource types, then open
them back up for updates in update_resource.  The danger is in JobSpec,
which can no longer create new resource types.  This also effectively
makes it a parse error for a user to specify an unknown resource, which
will happen as soon as the jobspec hits qmanager.

This changes the output of one test of resource_query from emitting
"unsatisfied" as info to emitting a descriptive parse error because it
was unsatisfied due to unknown resources.  That's why one test output
file is modified.